Extract out most of the tests into `tests/`
The only tests not extracted are compiler and parser,
which depend on the private `Token` type.
