+++ Cargo.toml
@@ -16,3 +16,7 @@ log = "0.3.5"
 
 [dev-dependencies]
 tempdir = "0.3.4"
+
+[[test]]
+name = "test"
+path = "tests/test.rs"
