+++ tests/test-data/base.foo.mustache
